+++ NoteNest/MainWindow.TaskEvents.cs
@@ -1,4 +1,0 @@
-namespace NoteNest;
-
-// Moved to NoteNest.Views.NoteNestWorkspaceView.TaskEvents.cs (v1.5.5)
-public partial class MainWindow { }